+++ packages/kitchen/src/components/Button/index.tsx
@@ -1,5 +1,6 @@
 import React from "react";
 import styled, { useTheme } from "styled-components";
+import withScale from "../../hoc/withScale";
 import { KitchenComponent, NormalSizes } from "../../types";
 import { AccentColors } from "../../types/theme";
 import convertRGBToRGBA from "../../utils/convertRGBToRGBA";
@@ -292,4 +293,4 @@ const Suffix = styled.span`
   color: inherit;
 `;
 
-export default Button;
+export default withScale(Button);