+++ packages/kitchen/src/native/index.ts
@@ -10,8 +10,10 @@ export * from "./types";
 export { default as themes, createTheme } from "./themes";
 
 // ./components
-export { default as KitchenProvider } from "./components/Provider";
-export type { KitchenProviderProps } from "./components/Provider";
+export {
+  default as KitchenProvider,
+  type KitchenProviderProps,
+} from "./components/Provider";
 
 export { default as Button, type ButtonProps } from "./components/Button";
 
@@ -19,6 +21,14 @@ export {
   default as Container,
   type ContainerProps,
 } from "./components/Container";
+
+export { default as Icon, type IconProps } from "./components/Icon";
+
+export {
+  default as Input,
+  type InputProps,
+  type InputChangeEventData,
+} from "./components/Input";
 
 export { default as Text, type TextProps } from "./components/Text";